Documented `GroupMe::User` class and revamped cmake
I have continued to document this library and have done `GroupMe::Self` and `GroupMe::User` so far. These all
should be generated using Doxygen and there should be a Github Action to generate the Docs and a Github Pages
set up to host the docs.

The CMake portion of this project has also been revamped to go up to standards with [This blog](https://blog.nickelp.ro/posts/cmake-pkg/)