Finished documenting main files
Finished documenting neccesary files and classes. The class contained in
include/base/multipart_parser.h needs to be looked at and improved. Lets add
that to the todo list. Other than that, all the classes that have currently been implemented
and that are neccesary to use this library have been documented with doxygen.