Started implementing a message class
I have started implementing a message class that will be used to hold message data,
for example a string of text, an attatchment, a sender, a time, etc.

I have also fixed some naming stuff, like renamed Attatchment.h to Attachment.h because
I spelt it wrong.

There is still a predicament that I need to figure out before I can continue. That is,
I want to not have to create the same user multiple times for different groups and messages,
so I can create a class that contains contacts. But how to get the contacts to everywhere...