Implemented the Message class
I implemented everything defined in the Message.h header file.
Some things may have been glossed over in the writing of the
header file so there needs to be a check on the header to see
if we have everything we need.

Also added the directory .cache/ to the .gitignore