Created custom unordered_set for storing Users
The new file UserSet.hpp includes a typedef called UserSet which is an unordered_set
that uses custom hash generation and custom comparison operations for the User class.

This might be the wrong type of set to use, as a regular set has a definate order, though
I am not sure if we need a definate order as of currently.