Basic implementation of the special query functions
I replaced the vector that holds the messages with a map with a key of
an `unsigned long long int` which holds the time the message was sent,
so the map will automatically sort the messages based on the time and I
don't have to worry about any algorithms.

Following that, I have also implemented basic implementation for the
query after, before, and since functions, and since I made the change to
a map I don't have to do anything special with them, just insert and
they will be sorted.

In addition to that I also changed the access for the `nlohmann::json`
objects from standard array access operators (e.g. []) to the `.at()`
function which has boundary detection.