Querying messages and started adding safety
Basic message querying has been added, I would like to improve this a
lot further though, I have to think about what needs improvement though.

I've also added safety to the constructor in the case errors occur when
creating the group via fetching it through the web. C++23 introduces a
type called std::expected which is similar to std::optional but has a
value for an error, considering moving to 23 to make use of that.