Added function to create `GroupMe::User` from a json
Seemed to simple to be true, but, alas, the functionality is here.

As with the `GroupMe::Message::createFromJson()` function you pass in a
json and it will construct a user, though for this all that is required
is a "user_id" element in the json, and everything else is optional, and
will be defaulted to "" if it is not set. It is **_HIGHLY_** reccomended
that you pass in those other options if possible. Although some of the4
options can not be applied to the user object, like "is_muted".