Fixed bug with the `GroupChat::removeGroupMember` function
If the user id didn't exist in the group, the function to grab the
relavent member group id would fail due to the member not being in the
group in the first place. There is now try and catch to catch that
exception and will now return a `BasicChat::Result::NotFound` if that
happens.