Added support for SubGroups (or topics)
This commit moves around a bunch of classes and reorginizes them to
better support class heiarchies. Each GroupChat class will have a
`std::list` of SubGroupChats that is constructed based on the return of
the server.

Each SubGroupChat has:
    - A name (topic)
    - A description (optional)
    - A parent ID
    - An ID
    - A Group Image (avatar)

Along with messages for the chat.

This commit also brings move and copy constructors to the classes.