Added permissions and join question and changed from using `std::shared_ptr` for `SubGroupChat`
The `GroupChat` class now has member variables and functions to
represent different permissions for multiple different things, such as
deletion of messages, ability to change `GroupChat` and `SubGroupChat`
settings, etc.

The ability to set a join question has also been implemented.

An instance of the `SubGroupChat` class stored in the `GroupChat` class
is only owned by the `GroupChat` class and *only* the `GroupChat` class,
therefor using a `std::shared_ptr` for it wouldn't be needed.

This might be rethought however, due to the `SubGroupChat` class being a
type of `GroupChat` class, and being able to exist on it's own.