Fixed interpretation of `type` and `visibility` in API Json Responses
The `type` entry in the API Responses were to distinguish an actual
group type. This however doesn't seem to be relavent after the creation
of the group, I believe this is to distinguish between announcement
groups and regular groups.

The `visibility` entry in the API Responses were to distinguish the
actual visibility of the group to people not already in the group.

This oversight has been rightfully corrected in the implementation.