Changed implementation for `SubGroupChat` using `std::unqiue_ptr`
Since a `SubGroupChat` is a type of `GroupChat`, it should be able to
stand independently like one too. There is nothing tying a `SubGroupChat`
to it's `GroupChat` owner (other than the couple of member variables
that must be synced to it), so it should be allowed to be independent
from it's GroupChat.

However, a `SubGroupChat` must need an *original* `GroupChat` owner, in
order to sync the permission data from the original owner `GroupChat`.
This is the only requirement.