Added ability to send messages
Each class derived from `GroupMe::BasicChat` will have a `sendMessage`
function that will allow sending a `GroupMe::Message` to the chat. Raw
text messages work, along with Picture and Video attachments, however
file uploading is broken for the moment.

In order to make this easier there is a new function in each class
derived from GroupMe::Attachment called `toJson` that will return a JSON
representation of the message in the form of a message that can be used
with GroupMe's servers.